Pocinka predupreždenij 32-bitnoj sborki.

+++ source/code_builder_lib/code_builder_types.hpp
@@ -685,7 +685,7 @@ struct Enum
 	Enum( const ProgramString& name, NamesScope* parent_scope );
 
 	NamesScope members;
-	uint64_t element_count= 0u;
+	size_t element_count= 0u;
 	FundamentalType underlaying_type; // must be integer
 
 	const Synt::Enum* syntax_element= nullptr; // Null if completed
